refactor(scripting): merge schemaCoverage.ts into testing.ts
Single file for all test infrastructure: mock registry, vi.mock
setup, schema walker, value generator, and assertSchemaCoverage.
Consumers import { mocks, assertSchemaCoverage } from one place.

Fixed consumer test to use createRollupGetCallValue with proper types
instead of sequencerInboxReadContract with as any cast.
